\name{arrayImpute}
\alias{arrayImpute}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ A Tool for Exploring missing pattern and comparing imputed values}
\description{
arrayImpute is a software to handle missing problem in microarray data
This software provides its own GUI to help biologists who are not familiar with R commands.
}
\usage{
arrayImpute()
}


\author{ Eun-kyung Lee, Dankyu Yoon, Taesung Park}
\examples{
arrayImpute()
}
\keyword{ dplot }% at least one, from doc/KEYWORDS
\keyword{ utilities }% __ONLY ONE__ keyword per line
